Changes struc field sizes and adds padding
Padding should be done between fields and at the end of the struct.
Each field is data-size aligned "struct is stored at an offset such that
offset mod data_size = 0"

Then each struct is paddded to align according to the biggest field
size. In our case game state needs to be aligned to an address that is
divideble by 8. The struct has the size 8+2+2=12 otherwise. No alignment
between fields is necessery as all fields is already aligned properly.

The player struct biggest field is 2 bytes and the struct is already
aligned that way so no need to do anything.

(http://www.necessaryandsufficient.net/2014/06/struct-padding-in-gcc/)